# Conflicts:
#	app/src/main/java/com/obpoo/gfsfabricsoftware/salesorder/ui/SOorderDetails.java

+++ app/src/main/java/com/obpoo/gfsfabricsoftware/salesorder/ui/SOorderDetails.java
@@ -106,10 +106,10 @@ public class SOorderDetails extends BaseActivity {
         desc_so_order_det.setText(getCompl);
         price_so_det.setText(item.getOrderTotal());
         dis_so_det.setText(item.getDiscount()+"%");
-        paid_so_det.setText(String.valueOf(Integer.valueOf(item.getPaybleAmount())-Integer.valueOf(item.getLeftover())));
+        paid_so_det.setText(String.valueOf(Double.valueOf(item.getPaybleAmount())-Double.valueOf(item.getLeftover())));
 
 
-        total_so_det.setText(String.valueOf( Integer.valueOf(item.getOrderTotal()) -( Integer.valueOf(item.getOrderTotal()) *( Integer.valueOf(item.getDiscount())/100))));
+        total_so_det.setText(String.valueOf( Double.valueOf(item.getOrderTotal()) -( Double.valueOf(item.getOrderTotal()) *( Double.valueOf(item.getDiscount())/100))));
 
         itemDetailsses = item.getItemDetails();
         SO_order_det_fab_adp adapter = new SO_order_det_fab_adp(itemDetailsses,SOorderDetails.this);
